requirements.txt
anyio==4.8.0
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
confluent-kafka==2.8.0
fastavro==1.10.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
